fix : change variable name fix : sepearate theme type
