+++ src/App.tsx
@@ -10,7 +10,7 @@ const StyledTopestDiv = styled.div`
 `;
 const Spacer = styled.div`
   //NavigationBar가 상단에 fixed될때 다른 컴포넌트를 가리는 문제르 해결하기 위한 공백 공간
-  height: 45px;
+  height: 50px;
 `;
 function App() {
   return (
