+++ src/theme/sementicPalette.ts
@@ -1,0 +1,32 @@
+export const sementicPalette = {
+  // 시맨틱 컬러
+  //시맨틱 컬러 -> ??
+  //브랜드 컬러
+  kakaoYellow: '#fee500', // 카카오 메인 노란색
+  kakaoYellowHover: '#ffea2e', // 카카오 노란색 호버 상태
+  kakaoYellowActive: '#d5c000', // 카카오 노란색 활성 상태
+  kakaoYellowPressed: '#d5c000', // 카카오 노란색 눌림 상태
+  kakaoBrown: '#654321', // 카카오 갈색
+  kakaoBrownPressed: '#432a12', // 카카오 갈색 눌림 상태
+
+  // 배경색
+  backgroundDefault: '#ffffff', // 기본 배경색
+  backgroundDisabled: '#f3f4f5', // 비활성 배경색
+  backgroundFill: '#f7f8f9', // 채움 배경색
+
+  // 텍스트 색상
+  textDefault: '#2a3038', // 기본 텍스트 색상
+  textSub: '#b0b3ba', // 보조 텍스트 색상
+  textDisabled: '#dcdee3', // 비활성 텍스트 색상
+  textPlaceholder: '#b0b3ba', // 플레이스홀더 텍스트 색상
+
+  // 테두리 색상
+  borderDefault: '#dcdee3', // 기본 테두리 색상
+  borderDisabled: '#eeeff1', // 비활성 테두리 색상
+
+  // 상태 색상
+  critical: '#fa342c', // 경고/에러 색상
+  criticalBackground: '#fdf0f0', // 경고/에러 배경 색상
+  info: '#217cf9', // 정보 색상
+  infoBackground: '#eff6ff', // 정보 배경 색상
+};
