+++ src/theme/theme.ts
@@ -1,0 +1,11 @@
+import { palette } from './palette';
+import { sementicPalette } from './sementicPalette';
+import { spacing } from './spacing';
+import { typography } from './typography';
+
+export const theme = {
+  palette,
+  sementicPalette,
+  typography,
+  spacing,
+};
