+++ src/theme/typography.ts
@@ -1,0 +1,84 @@
+export const typography = {
+  title1Bold: {
+    fontSize: '20px',
+    fontWeight: 700,
+    lineHeight: '27px',
+  },
+  title1Regular: {
+    fontSize: '20px',
+    fontWeight: 400,
+    lineHeight: '27px',
+  },
+  title2Bold: {
+    fontSize: '16px',
+    fontWeight: 700,
+    lineHeight: '24px',
+  },
+  title2Regular: {
+    fontSize: '16px',
+    fontWeight: 400,
+    lineHeight: '24px',
+  },
+
+  subtitle1Bold: {
+    fontSize: '16px',
+    fontWeight: 700,
+    lineHeight: '22px',
+  },
+  subtitle1Regular: {
+    fontSize: '16px',
+    fontWeight: 400,
+    lineHeight: '19px',
+  },
+  subtitle2Bold: {
+    fontSize: '14px',
+    fontWeight: 700,
+    lineHeight: '19px',
+  },
+  subtitle2Regular: {
+    fontSize: '14px',
+    fontWeight: 700,
+    lineHeight: '19px',
+  },
+
+  body1Bold: {
+    fontSize: '16px',
+    fontWeight: 700,
+    lineHeight: '22px',
+  },
+  body1Regular: {
+    fontSize: '16px',
+    fontWeight: 400,
+    lineHeight: '22px',
+  },
+  body2Bold: {
+    fontSize: '14px',
+    fontWeight: 700,
+    lineHeight: '19px',
+  },
+  body2Regular: {
+    fontSize: '14px',
+    fontWeight: 400,
+    lineHeight: '19px',
+  },
+  label1Bold: {
+    fontSize: '14px',
+    fontWeight: 700,
+    lineHeight: '19px',
+  },
+  label1Regular: {
+    fontSize: '14px',
+    fontWeight: 400,
+    lineHeight: '19px',
+  },
+  label2Bold: {
+    fontSize: '12px',
+    fontWeight: 700,
+    lineHeight: '16px',
+  },
+  label2Regular: {
+    fontSize: '12px',
+    fontWeight: 400,
+    lineHeight: '16px',
+  },
+};
